+++ tgbotapi.api/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/api/stickers/CreateNewVideoStickerSet.kt
@@ -1,0 +1,54 @@
+package dev.inmo.tgbotapi.extensions.api.stickers
+
+import dev.inmo.tgbotapi.bot.TelegramBot
+import dev.inmo.tgbotapi.requests.abstracts.FileId
+import dev.inmo.tgbotapi.requests.abstracts.MultipartFile
+import dev.inmo.tgbotapi.requests.stickers.CreateNewVideoStickerSet
+import dev.inmo.tgbotapi.types.CommonUser
+import dev.inmo.tgbotapi.types.UserId
+import dev.inmo.tgbotapi.types.stickers.MaskPosition
+
+suspend fun TelegramBot.createNewVideoStickerSet(
+    userId: UserId,
+    name: String,
+    sticker: FileId,
+    emojis: String,
+    containsMasks: Boolean? = null,
+    maskPosition: MaskPosition? = null
+) = execute(
+    CreateNewVideoStickerSet(userId, name, sticker, emojis, containsMasks, maskPosition)
+)
+
+suspend fun TelegramBot.createNewVideoStickerSet(
+    userId: UserId,
+    name: String,
+    sticker: MultipartFile,
+    emojis: String,
+    containsMasks: Boolean? = null,
+    maskPosition: MaskPosition? = null
+) = execute(
+    CreateNewVideoStickerSet(userId, name, sticker, emojis, containsMasks, maskPosition)
+)
+
+
+suspend fun TelegramBot.createNewVideoStickerSet(
+    user: CommonUser,
+    name: String,
+    sticker: FileId,
+    emojis: String,
+    containsMasks: Boolean? = null,
+    maskPosition: MaskPosition? = null
+) = createNewVideoStickerSet(
+    user.id, name, sticker, emojis, containsMasks, maskPosition
+)
+
+suspend fun TelegramBot.createNewVideoStickerSet(
+    user: CommonUser,
+    name: String,
+    sticker: MultipartFile,
+    emojis: String,
+    containsMasks: Boolean? = null,
+    maskPosition: MaskPosition? = null
+) = createNewVideoStickerSet(
+    user.id, name, sticker, emojis, containsMasks, maskPosition
+)